float to int

+++ Streamlit_Deployment/Streamlit_Final.py
@@ -200,7 +200,7 @@ st.dataframe(readmission_pivot)
 
 # Attempted to create a chart to show the comparison between the two main variables but cannot make it work on Streamlit
 readmission_hospital_nonull['effectiveness_of_care_national_comparison'] = readmission_hospital_nonull['effectiveness_of_care_national_comparison'].astype(str)
-readmission_hospital_nonull['readmission_national_comparison_footnote'] = readmission_hospital_nonull['readmission_national_comparison_footnote'].astype(float)
+readmission_hospital_nonull['readmission_national_comparison_footnote'] = readmission_hospital_nonull['readmission_national_comparison_footnote'].astype(int)
 df_Chart1 = readmission_hospital_nonull[['effectiveness_of_care_national_comparison', 'readmission_national_comparison_footnote']]
 st.bar_chart(df_Chart1)
